Got display working and no more crashes

+++ src/memory.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include "memory.h"
 
 char *memBanks[4],*sharedMem;
@@ -11,14 +12,16 @@ void initMemory() {
 	
 	sharedMem = malloc(SHARED_SIZE);
 }
-int myMMU(I8080 *cpu, int ladr) {
+
+
+int64_t myMMU(I8080 *cpu, int ladr) {
 	
-	int retval;
+	int64_t retval;
 	
 	if (bankNumber > 0) {
-		retval = ladr + (int) memBanks[bankNumber];
+		retval = ladr + (int64_t) memBanks[bankNumber];
 	} else if (ladr < 0x0800) {
-		retval = ladr + (int) memBanks[0];
+		retval = ladr + (int64_t) memBanks[0];
 	} else if (ladr < 0x5800) {
 		
 		// If it's in output mode then we know it's writing to the buffer
@@ -26,9 +29,9 @@ int myMMU(I8080 *cpu, int ladr) {
 			video.dirtyBuffer = 1;
 		}
 
-		retval = (int) video.pixelBuffer + ladr - 0x0800;
+		retval = (int64_t) video.pixelBuffer + ladr - 0x0800;
 	} else {
-		retval = ladr + (int) memBanks[0] - 0x5000;
+		retval = ladr + (int64_t) memBanks[0] - 0x5000;
 	}
 	
 	return retval;
